pull code out into separate modules; move files to /support
