+++ features/step_definitions/assertion_steps.rb
@@ -1,82 +1,9 @@
-module PathHelpers
-  # remove any leading locale path info
-  def current_path_without_locale(path)
-    locale_pattern =  /^(\/)(en|sv)?(\/)?(.*)$/
-    path.gsub(locale_pattern, '\1\4')
-  end
+require_relative File.join(__dir__, '..', 'support', 'step_errors')
 
-  def get_path(pagename, user = @user)
-    case pagename.downcase
-      when  'login'
-        path = new_user_session_path
-      when 'landing'
-        path = root_path
-      when 'new application'
-        path = new_shf_application_path
-      when 'edit application', 'edit my application'
-        user.reload
-        path = edit_shf_application_path(user.shf_application)
-      when 'application', 'show my application'
-        path = shf_application_path(user.shf_application)
-      when 'user instructions'
-        path = information_path
-      when 'member instructions'
-        path = information_path
-      when 'all waiting for info reasons'
-        path = admin_only_member_app_waiting_reasons_path
-      when 'new waiting for info reason'
-        path = new_admin_only_member_app_waiting_reason_path
-      when 'register as a new user'
-        path = new_user_registration_path
-      when 'edit registration for a user'
-        path = edit_user_registration_path
-      when 'new password'
-        path = new_user_password_path
-      when 'all member app waiting reasons'
-        path = admin_only_member_app_waiting_reasons_path
-      when 'business categories'
-        path = business_categories_path
-      when 'membership applications', 'shf applications'
-        path = shf_applications_path
-      when 'all companies'
-        path = companies_path
-      when 'create a new company'
-        path = new_company_path
-      when 'submit new membership application'
-        path = new_shf_application_path
-      when 'my first company'
-        path = company_path(user.shf_application.companies.first)
-      when 'my second company'
-        path = company_path(user.shf_application.companies.second)
-      when 'my third company'
-        path = company_path(user.shf_application.companies.third)
-      when 'edit my company'
-        path = edit_company_path(user.shf_application.companies.first)
-      when 'all users'
-        path = users_path
-      when 'all shf documents'
-        path = shf_documents_path
-      when 'new shf document'
-        path = new_shf_document_path
-      when 'user details', 'user profile'
-        path = user_path(user)
-      when 'test exception notifications'
-        path = test_exception_notifications_path
-      when 'admin dashboard'
-        path = admin_only_dashboard_path
-    when 'admin edit app configuration'
-      path = admin_only_edit_app_configuration_path
-      when 'admin show app configuration'
-        path = admin_only_app_configuration_path
-    end
 
-    expect(path).not_to be_empty, "A step was called with path= '#{pagename}', but that path is not defined in #{__method__} \n    (which is in #{__FILE__}"
+World(PathHelpers)  # include the PathHelpers module
 
-    path
-  end
-end
 
-World(PathHelpers)
 
 Then "I should{negate} see {capture_string}" do |negate, content|
   begin
