+++ features/step_definitions/navigation_steps.rb
@@ -1,20 +1,13 @@
+require_relative '../support/visit_page.rb'
+
+
+World(VisitPage)  # include the VisitPage module
+
+
+
 Given(/^I am on the "([^"]*)" page(?: for "([^"]*)")?$/) do |page, email|
-  user = email == nil ? @user :  User.find_by(email: email)
-
-  begin
-    visit path_with_locale(get_path(page, user))
-  rescue => exception
-    warn exception.message
-    begin
-      path_components = page.split(/\s+/)
-      visit self.send(path_components.push('path').join('_').to_sym)
-    rescue NoMethodError, ArgumentError
-      raise "Can't find mapping from \"#{page}\" to a path.\n" +
-        "Now, go and add a mapping in #{__FILE__} or assertion_steps.rb"
-    end
-  end
-
-
+  user = email == nil ? @user : User.find_by(email: email)
+  visit_page(page, user)
 end
